Add Debian config

+++ install.sh
@@ -5,41 +5,61 @@ read -p "Enter absolute path to dev directory [~/dev]: " dev_dir
 dev_dir=${dev_dir:-~/dev}
 mkdir -p $dev_dir
 
-# Clone dotfiles repository.
-[ -d $dev_dir/dotfiles ] || git clone https://github.com/nshki/dotfiles.git $dev_dir/dotfiles
+# OS selection.
+while [[ $os != "debian" && $os != "macos" ]]
+do
+  read -p "What OS are you using? [debian|mac] " os
+done
 
-# Ensure CLT for Xcode are installed.
-xcode-select --install
+# Install apps.
+if [[ $os == "debian" ]]; then
+  bash $dev_dir/dotfiles/config/debian/install_apps.sh
+elif [[ $os == "macos" ]]; then
+  # Ensure CLT for Xcode are installed.
+  xcode-select --install
 
-# Install Homebrew and bundle.
-/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-cd $dev_dir/dotfiles/config/macos/
-brew bundle
-cd $dev_dir
+  # Install Homebrew and bundle.
+  /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+  cd $dev_dir/dotfiles/config/macos/
+  brew bundle
+  cd $dev_dir
+fi
 
-# Symlink dotfiles.
-[ -f ~/.bash_local ] || touch ~/.bash_local
-rm -f ~/.bash_aliases && ln -s $dev_dir/dotfiles/config/.bash_aliases ~/.bash_aliases
-rm -f ~/.bash_profile && ln -s $dev_dir/dotfiles/config/macos/.bash_profile ~/.bash_profile
-rm -f ~/.gitignore_global && ln -s $dev_dir/dotfiles/config/.gitignore_global ~/.gitignore_global
-rm -f ~/.tmux.conf && ln -s $dev_dir/dotfiles/config/.tmux.conf ~/.tmux.conf
-rm -f ~/.tmuxline && ln -s $dev_dir/dotfiles/config/.tmuxline ~/.tmuxline
-rm -f ~/.vimrc && ln -s $dev_dir/dotfiles/config/.vimrc ~/.vimrc
+# Install tpm.
+git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 
 # Configure nvim to use .vimrc.
 mkdir -p ~/.config/nvim
 cp $dev_dir/dotfiles/config/init.vim ~/.config/nvim
 
-# Install tpm.
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+# Clone dotfiles repository.
+[ -d $dev_dir/dotfiles ] || git clone https://github.com/nshki/dotfiles.git $dev_dir/dotfiles
 
-# Add Terminal theme.
-open $dev_dir/dotfiles/config/macos/shikidark.terminal
+# Add Terminal themes.
+if [[ $os == "debian" ]]; then
+  bash -c "$(curl -fsSL https://raw.githubusercontent.com/denysdovhan/gnome-terminal-one/master/one-dark.sh)"
+elif [[ $os == "macos" ]]; then
+  open $dev_dir/dotfiles/config/macos/shikidark.terminal
+fi
+
+# Symlink and configure dotfiles.
+[ -f ~/.bash_local ] || touch ~/.bash_local
+rm -f ~/.bash_aliases && ln -s $dev_dir/dotfiles/config/.bash_aliases ~/.bash_aliases
+rm -f ~/.gitignore_global && ln -s $dev_dir/dotfiles/config/.gitignore_global ~/.gitignore_global
+rm -f ~/.tmux.conf && ln -s $dev_dir/dotfiles/config/.tmux.conf ~/.tmux.conf
+rm -f ~/.tmuxline && ln -s $dev_dir/dotfiles/config/.tmuxline ~/.tmuxline
+rm -f ~/.vimrc && ln -s $dev_dir/dotfiles/config/.vimrc ~/.vimrc
+if [[ $os == "macos" ]]; then
+  rm -f ~/.bash_profile && ln -s $dev_dir/dotfiles/config/macos/.bash_profile ~/.bash_profile
+elif [[ $os == "debian" ]]; then
+  echo -e "\nset -o vi" >> ~/.bashrc
+  echo -e '\nPATH="~/.rbenv/bin:$PATH"' >> ~/.bashrc
+fi
 
 # Setup Git.
 read -p "Enter Git name: " git_name
 read -p "Enter Git email: " git_email
-git config --global user.name $git_name
+git config --global user.name "$git_name"
 git config --global user.email $git_email
 git config --global core.excludesfile ~/.gitignore_global
 
@@ -51,7 +71,12 @@ if [ ! -f ~/.ssh/id_rsa ]; then
   ssh-keygen -t rsa -b 4096 -C $github_email
 fi
 eval "$(ssh-agent -s)"
-cp $dev_dir/dotfiles/config/sshconfig ~/.ssh/config
-ssh-add -K ~/.ssh/id_rsa
-pbcopy < ~/.ssh/id_rsa.pub
+if [[ $os == "macos" ]]; then
+  cp $dev_dir/dotfiles/config/sshconfig ~/.ssh/config
+  ssh-add -K ~/.ssh/id_rsa
+  pbcopy < ~/.ssh/id_rsa.pub
+elif [[ $os == "debian" ]]; then
+  eval "$(ssh-agent -s)"
+  xclip -sel clip < ~/.ssh/id_rsa.pub
+fi
 read -p "SSH key copied to clipboard. Add to GitHub (https://github.com/settings/keys). Press enter when done."
